requirements.txt
aiodataloader==0.4.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.13.0
amqp==5.3.1
aniso8601==7.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
APScheduler==3.10.4
asgiref==3.11.0
attrs==25.3.0
babel==2.17.0
backports.zstd==1.0.0
bcrypt==4.2.1
beautifulsoup4==4.14.2
bidict==0.23.1
billiard==4.2.2
blinker==1.9.0
Brotli==1.1.0
cachelib==0.9.0
cairocffi==1.7.1
captcha==0.7.1
celery==5.3.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.11.0
cryptography==44.0.3
cssselect2==0.8.0
curl_cffi==0.14.0
CurrencyConverter==0.17.10
Deprecated==1.2.18
distro==1.9.0
dj-database-url==3.0.1
Django==6.0
django-filter==25.2
django-jazzmin==3.0.1
django-leaflet==0.33.0
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
dnspython==2.7.0
docstring_parser==0.17.0
email-validator==2.1.0.post1
et_xmlfile==2.0.0
eventlet==0.40.2
ezdxf==1.4.3
Faker==20.1.0
fastapi==0.122.1
flake8==7.3.0
Flask==3.0.0
Flask-Admin==2.0.2
flask-babel==4.0.0
Flask-Caching==2.1.0
Flask-Compress==1.20
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
fonttools==4.55.0
forex-python==1.0.0
frozendict==2.4.7
frozenlist==1.7.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
gunicorn==21.2.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
kombu==5.5.4
limits==5.5.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
multidict==6.6.4
multitasking==0.0.12
mysqlclient==2.2.7
numpy==2.2.6
openai==2.2.0
openpyxl==3.1.2
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
peewee==3.18.3
pillow==10.4.0
platformdirs==4.5.1
pluggy==1.6.0
prometheus_client==0.23.1
prometheus_flask_exporter==0.23.2
promise==2.3
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.33.2
psutil==6.1.1
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pydyf==0.11.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pyphen==0.17.0
pypng==0.20220715.0
pyproj==3.7.2
pytest==8.4.2
pytest-cov==5.0.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.11.0
pytz==2025.2
qrcode==7.4.2
redis==5.0.1
reportlab==4.0.7
requests==2.31.0
rich==13.9.4
Rx==1.6.3
scikit-learn==1.7.2
scipy==1.15.3
sentry-sdk==2.35.0
setuptools==65.5.0
simple-websocket==1.1.0
simplejson==3.20.2
singledispatch==3.7.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sqlparse==0.5.4
starlette==0.50.0
tabulate==0.9.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tqdm==4.67.1
twilio==9.8.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
validators==0.28.3
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.14
weasyprint==62.3
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.0.1
wheel==0.45.1
whitenoise==6.11.0
wrapt==1.17.3
wsproto==1.2.0
WTForms==3.1.1
WTForms-SQLAlchemy==0.4.2
xlsxwriter==3.2.5
yarl==1.20.1
yfinance==0.2.66
zopfli==0.2.3.post1
zstandard==0.25.0
pymysql
